+++ graysvr/CContain.cpp
@@ -1523,7 +1523,7 @@ CChar * CItemCorpse::IsCorpseSleeping() const
 	// CItemCorpse
 	if ( !IsType(IT_CORPSE) )
 	{
-		DEBUG_ERR(("Corpse (0%x) doesn't have type T_CORPSE!(it has %d)\n",GetUID(),this->GetType()));
+		DEBUG_ERR(("Corpse (0%x) doesn't have type T_CORPSE!(it has %d)\n",(DWORD)GetUID(),this->GetType()));
 		return NULL;
 	}
